Non-struct-methods are capitalized.
This makes it clearer which functions are ‘methods’ on a given type (`cest__foo` is a ‘method’ on the `cest` type, which is a `struct cest*`), and which are functions stored under that type’s namespace (`Cest_bar` is the `bar()` function, stored under `Cest`—otherwise unrelated to the `cest` type).

This convention also makes it possible to programmatically determine which functions to treat as ‘methods’—that is, which to assume will always take a first argument of the same type as the start of the name, as `this`. (That is, `foo__bar()`, and thus `Foo.bar()`, will always take something of type `struct foo*` as its first argument)