No more signed `ll_ssize` or negative indicies.
Currently, you’re going to have to do the math yourself, if you want to access an element based on the end of a list (what negative indicies used to do). We may add a convenience function for this later, but it will always use positive integers; no more wasted space for signed integer versions of indicies.