(api new incomplete) Lots of new threading code.
This is very incomplete; it compiles, and the tests that are written pass… but none of the written tests really dig into the actual threading functionality, beyond creation and destruction. A much, much more robust and complex test suite needs to be built for this; don’t trust it yet.

Also, this doesn’t actually *do* anything as of yet, because `Routine.execute()` doesn’t exist; in fact, it might never exist, because I want to replace that with an ‘execution’ object that is sired by routines and then passed off to the `Thread` interface.