(incomplete minor fix) Ensuring `magazine`’s `bytes` array holds `unsigned char`s
Not sure if this is a comprehensive fix for the problem, though; there’s a lot of places where we use `char` instead of `unsigned char`. I tried to replace them all, but ran into cascading errors involving string literals (which are `char[]`, not `unsigned char[]`). Not sure how to go about this.